Создание прототипа
-Допилено обновление точки

+++ app/Http/Controllers/UpdatePointController.php
@@ -18,35 +18,47 @@ class UpdatePointController extends Controller
     {
         if (Auth::check()) {
             $validateFields = $request->validate([
-                'name' => ['max:255'],
-                'address' => ['max:255'],
-                'type' => ['max:255'],
-                'status' => ['max:255'],
+                'name' => ['string','required'],
+                'address' => ['string','required'],
+                'type' => ['string','required'],
+                'status' => ['string','required'],
                 'description' => ['max:255'],
                 'shortdescription' => ['max:255'],
-                'photo'  //УБРАТЬ ОГРАНИЧЕНИЕ
+                'photo'=> ['max:255']
+                //Допилить валидацию, как минимум, чтобы статус был только одной из возможных переменных
+
 
             ]);
+            if(isset($validateFields['photo'])){
+                $path = Storage::putFile('public/pointphoto', $request->file('photo'));
+                $path = 'storage/pointphoto/' . explode('/', $path)[2];
+                $oldpath = 'public/pointphoto/' . explode('/', $_SESSION['CurrentEditPoint']->photo)[2];
+                $delete = Storage::delete($oldpath);
+            }else{
+                $path = $_SESSION['CurrentEditPoint']->photo;
 
-            $path = Storage::putFile('public/pointphoto', $request->file('photo'));
-            $oldpath = 'public/pointphoto/' . explode('/', $_SESSION['CurrentEditPoint']->photo)[2];
+            }
 
-            $delete = Storage::delete($oldpath);
 
-            $path = 'storage/pointphoto/' . explode('/', $path)[2];
+
+
             $typeAndIcon = explode(',', $validateFields['type']);
             $validateFields['type'] = $typeAndIcon[1];
             $validateFields['icon'] = $typeAndIcon[0];
+
             DB::table('points')
                 ->where('id', $_SESSION['CurrentEditPoint']->id)
                 ->update([
                     'name' => $validateFields['name'],
-                    'photo' => $path, 'status' => $validateFields['status'],
-                    'address' => $validateFields['address'],
                     'type' => $validateFields['type'],
                     'icon' => $validateFields['icon'],
+                    'address' => $validateFields['address'],
+                    'status' => $validateFields['status'],
                     'description' => $validateFields['description'],
-                    'shortdescription'=>$validateFields['shortdescription'],
+                    'shortdescription' => $validateFields['shortdescription'],
+                    'photo' => $path,
+
+
                 ]);
             return redirect(route('GetUpdatePoint', $_SESSION['CurrentEditPoint']->id));
         }
@@ -55,14 +67,27 @@ class UpdatePointController extends Controller
 
     public function GetUpdatePoint($id)
     {
-        $id = (int)$id;
-        if ((is_numeric($id)) and ($id > 0)) {
-            $_SESSION['CurrentEditPoint'] = DB::table('points')->select('id', 'name', 'type', 'icon', 'address', 'description', 'shortdescription', 'status', 'photo')
-                ->where('id', $id)->first();
-            $_SESSION['CurrentEditPoint']->type = $_SESSION['CurrentEditPoint']->icon . ',' . $_SESSION['CurrentEditPoint']->type;
-            return view('editpoints');
+        if (Auth::check()) {
+            $id = (int)$id;
+            if ((is_numeric($id)) and ($id > 0)) {
+                $_SESSION['CurrentEditPoint'] = DB::table('points')
+                    ->select(
+                        'id',
+                        'name',
+                        'type',
+                        'icon',
+                        'address',
+                        'status',
+                        'description',
+                        'shortdescription',
+                        'photo'
+                    )
+                    ->where('id', $id)->first();
+                $_SESSION['CurrentEditPoint']->type = $_SESSION['CurrentEditPoint']->icon . ',' . $_SESSION['CurrentEditPoint']->type;
+                return view('editpoints');
+            }
         }
-
+        return redirect(route('login'));
 
     }
